add OrdersFilterBar tests.
